requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
python-multipart==0.0.6
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
sqlalchemy==2.0.20
alembic==1.12.0
openai==1.61.0
python-dotenv==1.0.1
transformers==4.48.2
huggingface-hub==0.28.1
safetensors==0.5.2
accelerate==1.3.0
urllib3==1.26.16
yt-dlp
openai-whisper
google-genai
websockets==11.0.3
psycopg2-binary
starlette==0.27.0